refactor: Rename some class with better names
