+++ src/dotnet/ReSharperPlugin.SpecflowRiderPlugin/QuickFixes/CreateMissingStepQuickFix.cs
@@ -4,6 +4,7 @@ using JetBrains.ReSharper.Feature.Services.QuickFixes;
 using JetBrains.ReSharper.Feature.Services.Resources;
 using JetBrains.Util;
 using ReSharperPlugin.SpecflowRiderPlugin.Daemon;
+using ReSharperPlugin.SpecflowRiderPlugin.Daemon.UnresolvedReferenceHighlight;
 using ReSharperPlugin.SpecflowRiderPlugin.QuickFixes.Actions;
 
 namespace ReSharperPlugin.SpecflowRiderPlugin.QuickFixes
